changed color cycle, added legend
color cycle is changed all around to make me like it more.
I added light gray as a color in the colors.py file.
I made a function that makes a pretty legend. It’s not totally done,
but is definitely functional. The thing that isn’t done is something I
have never actually used before, so it’s not urgent.