Added some improvements to scatter and hist
Scatter: added a function to calculate a roughly appropriate alpha
given the length of the dataset.
Hist: Added an automatic bin size calculator that is used if the user
doesn’t specify bins or bins size. This is based on the Freedman
Diaconis rule, but the bin size is rounded so that the edges line up
with ticks on the axes. This makes it look nicer to my eye.