use matplotlib objects
I am now making subclasses of the matplotlib objects. Then all the
functions I have defined are defined on those objects. There are still
some holes, and the documentation hasn’t been touched yet, but the
general picture is in shape, and it should all work. This should
provide a much better interface to the user, if it all works like I
think it will.