updates to scatter and legend for matplotlib 2.0
The updates to matplotlib 2.0 changed a few style things, so I have
adjusted things as well. Nothing will change if still on matplotlib
1.x.

Changes: edge color isn’t messed with in 2.0, since the borders are the
color of the points, which is nicer. The behavior in 1.x is adjusted to
match that now too, rather than just having black edges.
- The legend now has a translucent white background, rather than no
background. This is true for both 2.x and 1.x.
- The line width of the legend can now be adjusted, too, rather than
just being zero.
- The default color cycle is only adjusted for 1.x.